requirements.txt
pandas>=2.3
openpyxl>=3.1
PySide>=6.9